[AutoDiff] Remove the method mode from 'Builtin.autodiffApply'. (#22859)
Before indirect passing was supported, the `_method` configuration in `Builtin.autodiffApply_vjp_method` was used to directly apply a method's VJP as if its arguments were flattened like SIL. Not only was it a workaround, as a trade-off, it also introduced some semantic incorrectness in `@differentiable` functions: It made `@differentiable (T) -> (U) -> V` mean that `U` was also a differentiation parameter in the function, but `@differentiable` should not apply to inner curry levels.

After indirect passing support landed, we removed the method-specific differential operator from the standard library (#22637). So the `_method` mode in `Builtin.autodiffApply` becomes dead code. This PR removes that and simplifies the infrastructure.