[Sema] TF-288: Handle immutable properties in `Differentiable` derive… (#22905)
Previously, `Differentiable` derived conformances was not enabled for structs
with `let` stored properties with initial values. (This is because immutable
stored properties can never be updated.)

Now, immutable stored properties are implicitly marked with `@noDerivative`
and are diagnosed with a warning+fixit.

Resolves TF-288.

* Update diagnostic message.

Clarify that `let` stored properties with a default value can be changed
to `var` to allow derivatives.