[AutoDiff] Use owned callee convention for linear maps.
Switch to `@callee_owned` callee convention for all linear map functions (differentials and pullbacks) returned from derivative functions. This reduces a half of reference counting operations in compiler-generated derivatives, and enables child linear maps that are called in linear maps to be destroyed right after the call.

Background: Before this patch, linear map functions took an `@owned` context and had `@callee_guaranteed` callee convention. It was a suboptimal design because
1. All linear maps in AD-generated code are called exactly once. The context's parameter convention is `@owned` because we want to consume the context as early as possible, but doing so in combination with `@calee_guaranteed` convention leads to an unnecessary pair of retain (in the partial application forwarder) and release (in the caller). As a result, the entire context was kept alive until the entire outer pullback returns.
2. Pullbacks' allocation and deallocation follow a strict stack discipline, so we really want to consume pullbacks as early as possible and not retain unused memory.

Resolves rdar://71892494.